Added pattern matching capability for link and bearer options.
This patch allows certain link-related (or bearer-related) options
to operate on multiple links (or bearers) at a time.  To request
this capability, the user specifies a pattern consisting of a '?'
followed by a string of characters; all links (or bearers) whose name
contains the specified string are considered a match for the pattern.
A pattern consisting of a '?' alone matches all links (or bearers).

Signed-off-by: Allan Stephens <allan.stephens@windriver.com>