tipcutils: Fix a precedence bug and gcc warning
When compiling with gcc 4.7.2, this warning is thrown:
	tipc-pipe.c: In function ‘pipe_start’:
	tipc-pipe.c:290:3: warning: suggest parentheses around ‘&&’ within ‘||’

This patch just fixes it by adding the parentheses in a way that if either of
the fds receive a POLLHUP and no data_in_len, then it breaks and exits the
function.

Without parentheses "a || b && !c" is evaluated as "a || (b && !c)", so without
parentheses it will be true if "a" is true, no matter "c". And IIUC that is not
what we want, as we want to make sure it exits when no data is received.

Now it seems to work "by chance" and the bug is never hit. Because if we close
then "a" (the local POLLHUP) is true, and then the whole if is true (although
"c" is 0 in that case too) and exits. And when the other end closes, then "b" is
true (the other end POLLHUP) and "c" is 0 in that case (as it is initialized to
0 at the beggining of the loop) so b && !c is true, and the whole if is true.

But it seems that the right thing is to check a POLLHUP on any of the fds and no
data_in_len, as that is what we want to know. And on those cases will continue
to work with this change, because when "a" is true, "c" is 0 (so "(a || b) &&
!c" will be true) and when "b" is true, "c" is 0 too and the if will be true
too.